Project status auto-update to 'archived'.
Add functionality for Project to auto change it's status to 'archived'
when it has an archive relationship created. This saves the user from
having to update the project status on their own (likely forgetting to
do so).

The most common use case is for a project to be changed to
archived once it has an archive relationship. The user can still change
the project status back to something other than archived for
fringe/exception cases.