Generate barcode for drive
Show page gnerates a unique barcode based off the objects class name and
id. In this case a Drive. But we could extend this functionality to
other objects later if we wanted to. Barcode displays as a png generated
on the client side from a string via the barby gem.