Fix pagination by changing app root
In our production environment our app is scoped to a sub URI because it
shares the base URI with other apps and processes. This was breaking
pagination in production only. By moving the root route within our named
scope we ensure will_paginate will include it in the links it generates.
One trade off is in development the "/" route will take us to the
welcome aboard rails page. But we never call this in production because
nginx would just send us to the landing page for all the apps.