Use render over redirect for better security
In production we were getting a javascript security warning for
ReplaceState in History. The cause of this was my nginx conf violating
the new "single orgin" security policy for web browsers. The proxy pass
and the request did not match. Instead of making my nginx conf more
complex, we can call render here instead of redirect. That way we are
not changing the request at all and instead explicitly telling which
controller and action, in this case sessions and new, that we want to
execute.