Add print view
Each drive now has a print label link that launches a new tab in the
browser and diplays a small label from printing to a DYMO Label printer.
The Barcodes have chenged to QR codes in favor of having a standard
shorter width. QR codes can be scanned to quick launch the record for
tht drive. This will facilitate a check-in, check-out system.