Add ability to view each checkout record and edit
This edit action is separate from the return feature. We want those two
forms to be different to readily respond to the context of the user's
intent. Returning should be as quick, and easy as possible with fields
that auto fill. We've moved this to a separate checkins controller. Edit
should have all fields available for edit with no autofill other than
what's already saved in the databse.