Send email notification on drive create
On drive creation, sends an email to a list of recipients managed though
envorinment variables (for now) because our user accounts do not have an
email addess associated with them by default. They are auto generated
against active directory. So this is something we will have to have
users add on their own later on.