Display date created for drives
By request, users can now see when a record for each drive was created.
This is on the show page and on the index page. On the index we have
replaced the 'view' button with a date created column. Now the entire
row is clickable rather than just the drive name, so we no longer need a
redundant button. Also some minor styling changes to make things more
readable and neat.